% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BigCity}
\alias{BigCity}
\title{Full Person-level Population Database}
\format{An object of class \code{data.frame} with 150266 rows and 12 columns.}
\source{
\url{https://CRAN.R-project.org/package=TeachingSampling}
}
\usage{
data(BigCity)
}
\description{
This data set corresponds to some socioeconomic variables from 150266 people of a city in a particular year.
}
\references{
Package ‘TeachingSampling’; see \code{\link[TeachingSampling]{BigCity}}
}
\keyword{datasets}
