% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{C}
\alias{C}
\title{calculates the C factor}
\usage{
C(ys, xs, R, Vs)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}
}
\description{
calculates the C factor
}
