% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create1.R
\name{create1}
\alias{create1}
\title{creates vector of 1's to be used in the estimators}
\usage{
create1(y)
}
\arguments{
\item{y}{sample matrix}
}
\value{
vector of 1's with size equal to the number of observations in the sample
}
\description{
creates vector of 1's to be used in the estimators
}
