% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{E_beta}
\alias{E_beta}
\title{calculates the BLE for Beta}
\usage{
E_beta(ys, xs, a, R, Vs)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{a}{vector of means from Beta}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}
}
\description{
calculates the BLE for Beta
}
