% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{E_theta_Reg}
\alias{E_theta_Reg}
\title{calculates the BLE for the individuals not in the sample}
\usage{
E_theta_Reg(ys, xs, a, R, Vs, x_nots)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{a}{vector of means from Beta}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}

\item{x_nots}{values of X for the individuals not in the sample}
}
\description{
calculates the BLE for the individuals not in the sample
}
