% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{T_Reg}
\alias{T_Reg}
\title{calculates BLE for the total T}
\usage{
T_Reg(ys, xs, a, R, Vs, x_nots)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{a}{vector of means from Beta}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}

\item{x_nots}{values of X for the individuals not in the sample}
}
\description{
calculates BLE for the total T
}
