% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{V_beta}
\alias{V_beta}
\title{calculates the risk matrix associated with the BLE for Beta}
\usage{
V_beta(ys, xs, R, Vs)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}
}
\description{
calculates the risk matrix associated with the BLE for Beta
}
