% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{V_theta_Reg}
\alias{V_theta_Reg}
\title{calculates the risk matrix associated with the BLE for the individuals not in the sample}
\usage{
V_theta_Reg(ys, xs, R, Vs, x_nots, V_nots)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}

\item{x_nots}{values of X for the individuals not in the sample}

\item{V_nots}{covariance matrix of the individuals not in the sample}
}
\description{
calculates the risk matrix associated with the BLE for the individuals not in the sample
}
