% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLE_Reg.R
\name{VT_Reg}
\alias{VT_Reg}
\title{calculates risk matrix associated with the BLE for for the total T}
\usage{
VT_Reg(ys, xs, a, R, Vs, x_nots, V_nots)
}
\arguments{
\item{ys}{response variable of the sample}

\item{xs}{explicative variable of the sample}

\item{a}{vector of means from Beta}

\item{R}{covariance matrix of Beta}

\item{Vs}{covariance of sample errors}

\item{x_nots}{values of X for the individuals not in the sample}

\item{V_nots}{covariance matrix of the individuals not in the sample}
}
\description{
calculates risk matrix associated with the BLE for for the total T
}
